+++ testapp/src/main/java/com/rakuten/tech/mobile/testapp/helper/AppPermission.kt
@@ -1,7 +1,7 @@
 package com.rakuten.tech.mobile.testapp.helper
 
 import android.Manifest
-import com.rakuten.tech.mobile.miniapp.permission.MiniAppPermissionType
+import com.rakuten.tech.mobile.miniapp.permission.MiniAppDevicePermissionType
 
 object AppPermission {
 
@@ -9,13 +9,14 @@ object AppPermission {
         const val LOCATION = 1001
     }
 
-    fun getRequestCode(permissionType: MiniAppPermissionType) = when (permissionType) {
-        MiniAppPermissionType.LOCATION -> ReqCode.LOCATION
+    fun getDeviceRequestCode(permissionType: MiniAppDevicePermissionType) = when (permissionType) {
+        MiniAppDevicePermissionType.LOCATION -> ReqCode.LOCATION
         else -> 0
     }
 
-    fun getPermissionRequest(permissionType: MiniAppPermissionType) = when (permissionType) {
-        MiniAppPermissionType.LOCATION -> arrayOf(Manifest.permission.ACCESS_FINE_LOCATION)
-        else -> emptyArray()
-    }
+    fun getDevicePermissionRequest(permissionType: MiniAppDevicePermissionType) =
+        when (permissionType) {
+            MiniAppDevicePermissionType.LOCATION -> arrayOf(Manifest.permission.ACCESS_FINE_LOCATION)
+            else -> emptyArray()
+        }
 }